Add files via upload
Python code for my capstone Project. This project analyzes graduates from the University of Findlay to identify factors contributing to securing a job with an annual salary of $70,000 or more. The study begins with data preprocessing to prepare the data set for analysis. Following this, three machine learning models—Binary Logistic Regression, Decision Tree, and Random Forest—are employed and compared to determine which model performs best.